Add eimd-2019
We include a dynamic requirement on Apache POI (via 'docjure') which uses log4j2. The POI library logs an error during processing of the xlsx file published for England, so we add a custom logging configuration used only for that step to suppress that log output. In usual use, such as running a server, we do not include Apache POI and use slf4j instead as sqlite-jdbc includes a dependency on slf4j-api.